Fix: 32 bit compatibility

+++ src/lib/api.c
@@ -264,7 +264,7 @@ bool push_response(struct t_work_response *response) {
             MYMPD_LOG_DEBUG(NULL, "Push response to webserver queue for connection %lu: %s", response->conn_id, response->data);
             return mympd_queue_push(web_server_queue, response, 0);
         case RESPONSE_TYPE_RAW:
-            MYMPD_LOG_DEBUG(NULL, "Push raw response to webserver queue for connection %lu with %lu bytes", response->conn_id, sdslen(response->data));
+            MYMPD_LOG_DEBUG(NULL, "Push raw response to webserver queue for connection %lu with %lu bytes", response->conn_id, (unsigned long)sdslen(response->data));
             return mympd_queue_push(web_server_queue, response, 0);
         case RESPONSE_TYPE_SCRIPT:
             #ifdef MYMPD_ENABLE_LUA
